Suggestion for using 'with open( )' as syntax
I've changed the file handling code to use the recommended (in the python documentation) method. This form will gracefully close the file connection if something goes wrong (i.e. it doesn't need a seperate close() function and will close the connection if an error occurs). It also reads very nicely.